# Reduce iterations required for dead argument elimination convergence