Reduce iterations required for dead argument elimination convergence

Crude measurements 
Before:
```
[PassRunner]   running pass: dae-optimizing...                     53.1484 seconds.
```
After:
```
[PassRunner]   running pass: dae-optimizing...                     39.9249 seconds.
```



<details>
<summary>More details</summary>
<details>
<summary>Before</summary>

```
[PassRunner] running passes
[PassRunner]   running pass: dae-optimizing...                     
Function Count: 132035

Iteration: 1
  removedParams: 12961
  returnValuesRemoved: 0
Iteration: 2
  removedParams: 2343
  returnValuesRemoved: 0
Iteration: 3
  removedParams: 903
  returnValuesRemoved: 0
Iteration: 4
  removedParams: 553
  returnValuesRemoved: 0
Iteration: 5
  removedParams: 409
  returnValuesRemoved: 0
Iteration: 6
  removedParams: 303
  returnValuesRemoved: 0
Iteration: 7
  removedParams: 124
  returnValuesRemoved: 0
Iteration: 8
  removedParams: 100
  returnValuesRemoved: 0
Iteration: 9
  removedParams: 88
  returnValuesRemoved: 0
Iteration: 10
  removedParams: 69
  returnValuesRemoved: 0
Iteration: 11
  removedParams: 53
  returnValuesRemoved: 0
Iteration: 12
  removedParams: 41
  returnValuesRemoved: 0
Iteration: 13
  removedParams: 21
  returnValuesRemoved: 0
Iteration: 14
  removedParams: 17
  returnValuesRemoved: 0
Iteration: 15
  removedParams: 12
  returnValuesRemoved: 0
Iteration: 16
  removedParams: 10
  returnValuesRemoved: 0
Iteration: 17
  removedParams: 7
  returnValuesRemoved: 0
Iteration: 18
  removedParams: 4
  returnValuesRemoved: 0
Iteration: 19
  removedParams: 4
  returnValuesRemoved: 0
Iteration: 20
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 21
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 22
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 23
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 24
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 25
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 26
  removedParams: 0
  returnValuesRemoved: 16824
Iteration: 27
  removedParams: 11
  returnValuesRemoved: 0
Iteration: 28
  removedParams: 1
  returnValuesRemoved: 0
Iteration: 29
  removedParams: 0
  returnValuesRemoved: 1523
Iteration: 30
  removedParams: 1
  returnValuesRemoved: 0
Iteration: 31
  removedParams: 0
  returnValuesRemoved: 101
Iteration: 32
  removedParams: 0
  returnValuesRemoved: 73
Iteration: 33
  removedParams: 0
  returnValuesRemoved: 24
Iteration: 34
  removedParams: 0
  returnValuesRemoved: 2
Iteration: 35
  removedParams: 0
  returnValuesRemoved: 1

Total Iterations: 35

Total removedParams: 18047

Total returnValuesRemoved : 18548
53.1484 seconds.
```

</details>
<details>
<summary>After</summary>

```
[PassRunner] running passes
[PassRunner]   running pass: dae-optimizing...                     
Function Count: 132035

Iteration: 1
  removedParams: 12961
  returnValuesRemoved: 15818
Iteration: 2
  removedParams: 2352
  returnValuesRemoved: 2472
Iteration: 3
  removedParams: 906
  returnValuesRemoved: 154
Iteration: 4
  removedParams: 553
  returnValuesRemoved: 78
Iteration: 5
  removedParams: 409
  returnValuesRemoved: 27
Iteration: 6
  removedParams: 303
  returnValuesRemoved: 2
Iteration: 7
  removedParams: 124
  returnValuesRemoved: 1
Iteration: 8
  removedParams: 100
  returnValuesRemoved: 0
Iteration: 9
  removedParams: 88
  returnValuesRemoved: 0
Iteration: 10
  removedParams: 69
  returnValuesRemoved: 0
Iteration: 11
  removedParams: 53
  returnValuesRemoved: 0
Iteration: 12
  removedParams: 41
  returnValuesRemoved: 0
Iteration: 13
  removedParams: 21
  returnValuesRemoved: 0
Iteration: 14
  removedParams: 17
  returnValuesRemoved: 0
Iteration: 15
  removedParams: 12
  returnValuesRemoved: 0
Iteration: 16
  removedParams: 10
  returnValuesRemoved: 0
Iteration: 17
  removedParams: 7
  returnValuesRemoved: 0
Iteration: 18
  removedParams: 4
  returnValuesRemoved: 0
Iteration: 19
  removedParams: 4
  returnValuesRemoved: 0
Iteration: 20
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 21
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 22
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 23
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 24
  removedParams: 2
  returnValuesRemoved: 0
Iteration: 25
  removedParams: 2
  returnValuesRemoved: 0

Total Iterations: 25

Total removedParams: 18046

Total returnValuesRemoved : 18552
39.9249 seconds.
```

</details>

</details>