@sbc100 2022-05-05
This change is currently breaking the binaryen roll into emscripten.

I bisected the test failure to this PR:

```
$ ./tests/runner cores.test_getcwd_with_non_ascii_name
....
RuntimeError: Aborted(Assertion failed: longStr.length() == 255, at: /usr/local/google/home/sbc/dev/wasm/emscripten/tests/fs/test_getcwd_with_non_ascii_name.cpp,70,main). Build with -sASSERTIONS for more info.
    at abort (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:7906)
    at ___assert_fail (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:11285)
    at k (<anonymous>:wasm-function[52]:0x3aa0)
    at Module._main (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:60302)
    at callMain (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:60688)
    at doRun (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:61107)
    at run (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:61279)
    at runCaller (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:60543)
    at removeRunDependency (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:7710)
    at receiveInstance (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:9420)
    at receiveInstantiationResult (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:9541)
Thrown at:
    at abort (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:7906)
    at ___assert_fail (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:11285)
    at k (:1:15009)
    at Module._main (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:60302)
    at callMain (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:60688)
    at doRun (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:61107)
    at run (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:61279)
    at runCaller (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:60543)
    at removeRunDependency (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:7710)
    at receiveInstance (/tmp/emscripten_test_cores_xgqdan7k/test_getcwd_with_non_ascii_name.js:1:9420)
```

Only occurs with `-Os` not `-O2`.  Seems to change program semantics in such as way that the assertion in the test program fails.