@sbc100 2022-05-05
If you would prefer to fix forward rather than revert the current breakage is not actually hurting emscripten yet, its just preventing new changes from binaryen making their way to emscripten users.   Just FYI.