@arsnyder16 2022-05-06
@sbc100 I am struggling to reproduce. 
Here is what i am trying.  `test.baseline.wasm` and  `test.new.wasm` are binary matches 
```bash

em++ ../../emscripten/tests/fs/test_getcwd_with_non_ascii_name.cpp -O0 -o test.js
node test.js
mv test.wasm test.og.wasm

/root/emsdk/upstream/bin/wasm-opt -Os \
 --strip-dwarf --post-emscripten --low-memory-unused --zero-filled-memory --strip-debug --strip-producers \
 test.og.wasm -o test.wasm --mvp-features
node test.js
mv test.wasm test.baseline.wasm

bin/wasm-opt -Os \
 --strip-dwarf --post-emscripten --low-memory-unused --zero-filled-memory --strip-debug --strip-producers \
 test.og.wasm -o test.wasm --mvp-features
node test.js
mv test.wasm test.new.wasm

```

```
~/binaryen/build (asnyder/dae-iterations2)# ./test.sh
positive test case: абвгд
ret2 = /абвгд
positive test case: abcdeабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгд
ret2 = /abcdeабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгд
negative test case: abcdeabcdeabcdeabcdeabcdeabcdeabcdeabcdeabcdeabcde
positive test case: абвгд
ret2 = /абвгд
positive test case: abcdeабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгд
ret2 = /abcdeабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгд
negative test case: abcdeabcdeabcdeabcdeabcdeabcdeabcdeabcdeabcdeabcde
positive test case: абвгд
ret2 = /абвгд
positive test case: abcdeабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгд
ret2 = /abcdeабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгдабвгд
negative test case: abcdeabcdeabcdeabcdeabcdeabcdeabcdeabcdeabcdeabcde
```