@sbc100 2022-05-06
Did you try running exact test command from above `./tests/runner cores.test_getcwd_with_non_ascii_name` ?