@arsnyder16 2022-05-06
I don't have the full emscripten dev environment setup, also with it so i can integrate local binaryen.  

But i now am able to repro but swapping my local build of wasm-opt into my emsdk upstream then just running

```
em++ ../emscripten/tests/fs/test_getcwd_with_non_ascii_name.cpp -Os -o test.js
node test.js
```

Thanks! I'll get to the bottom of this