@sbc100 2022-05-06
emscripten ships with the test runner and test suite.   Just run `./tests/runner` from either and emscirpten checkout or the emskd version of emscipten (which em++ to see where that is).

You can configure emscripten to use your custom binaryen version using `EM_BINARYEN_ROOT` in your environment.